refactor(adapter-openclaw): move CLI-command error helpers into errors.ts
Second cut of the runtime.ts god-file split. Consolidates the OpenClaw CLI
command-failure error handling into the existing errors.ts (already "the ONE
place OpenClaw provider error strings are interpreted"): the
OpenClawCommandError class, its processOutput/processExitCode constructor
helpers, openClawErrorOutput, isPluginAllowlistOpenFailure, and the
OPENCLAW_PLUGIN_ALLOWLIST_WARNING sentinel.

errors.ts imports stripAnsi from runtime-utils (no cycle — runtime-utils is
the dependency-free leaf). runtime.ts imports back only the two symbols it
still references (OpenClawCommandError for throw/instanceof,
isPluginAllowlistOpenFailure); processOutput/processExitCode/
openClawErrorOutput were only used by the moved cluster, so they stay
module-internal to errors.ts.

Pure relocation — bodies byte-for-byte unchanged. runtime.ts 3,053 -> 2,989.
Verified: typecheck, eslint (changed files), adapter suite 115-0 (--isolate),
full suite 5124-0, binary build (stamps reverted), boot smoke (adapter loads,
schedule activates, GET /api/plugins/schedule/ -> 200).
